Use a different path for cygwin cache
Users dont need cygwin cache post installation. This is why we delete
them in postinstall.js. However, it's nice to retain them on CI for
faster Github workflows.

We,
1. Use a different path to store the downloaded cygwin packages (.cygwin-cache)
2. Copy them to npm tarball area (.cygwin/var/cache/setup)
3. Proceed with current flow and delete the cache from tarball
area (.cygwin/var/cache/setup) on client side.
4. .cygwin-cache remains intact on CI.
