+++ install-cygwin-packages.js
@@ -1,7 +1,8 @@
 const { installPackages } = require("./lib");
+const { localPackageDirectory } = require("./paths.js");
 
 if (process.platform === "win32") {
-  installPackages();
+  installPackages(localPackageDirectory);
 } else {
   console.log("Not Windows; Cygwin is not required.");
 }
